Simpler, safer LRU replacement
